+++ src/ast/variables/LocalVariable.ts
@@ -21,13 +21,16 @@ const MAX_PATH_DEPTH = 7;
 
 export default class LocalVariable extends Variable {
 	declarations: (Identifier | ExportDefaultDeclaration)[];
-	init: ExpressionEntity;
-	reassignmentTracker: EntityPathTracker;
+	init: ExpressionEntity | null;
+	isLocal: true;
+	additionalInitializers: ExpressionEntity[] | null = null;
+
+	private reassignmentTracker: EntityPathTracker;
 
 	constructor(
 		name: string,
 		declarator: Identifier | ExportDefaultDeclaration | null,
-		init: ExpressionEntity,
+		init: ExpressionEntity | null,
 		reassignmentTracker: EntityPathTracker
 	) {
 		super(name);
@@ -36,8 +39,25 @@ export default class LocalVariable extends Variable {
 		this.reassignmentTracker = reassignmentTracker;
 	}
 
-	addDeclaration(identifier: Identifier) {
+	addDeclaration(identifier: Identifier, init: ExpressionEntity | null) {
 		this.declarations.push(identifier);
+		if (this.additionalInitializers === null) {
+			this.additionalInitializers = this.init === null ? [] : [this.init];
+			this.init = UNKNOWN_EXPRESSION;
+			this.isReassigned = true;
+		}
+		if (init !== null) {
+			this.additionalInitializers.push(init);
+		}
+	}
+
+	consolidateInitializers() {
+		if (this.additionalInitializers !== null) {
+			for (const initializer of this.additionalInitializers) {
+				initializer.reassignPath(UNKNOWN_PATH);
+			}
+			this.additionalInitializers = null;
+		}
 	}
 
 	getLiteralValueAtPath(
@@ -151,3 +171,5 @@ export default class LocalVariable extends Variable {
 		}
 	}
 }
+
+LocalVariable.prototype.isLocal = true;