+++ test/function/samples/reassign-array-literal-elements/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'makes sure reassignments of array elements are tracked'
+};