+++ test/function/samples/reassign-pattern-defaults/_config.js
@@ -1,0 +1,4 @@
+module.exports = {
+	description: 'makes sure reassignments of pattern defaults are tracked',
+	minNodeVersion: 6
+};